requirements.txt
datasets
diffusers
fire
gradio
h5py
huggingface_hub
imageio
jaxtyping
matplotlib
numpy
opencv_python
pandas
Pillow
plyfile
git+https://github.com/facebookresearch/pytorch3d.git@stable
scipy
scikit-image
submitit
torch
torchvision
tqdm
trimesh
wandb
xformers
timm
transformers
accelerate
easydict
submodules/diff-gaussian-rasterization
submodules/simple-knn